listingsvc/listingsvc/requirements.txt
asgiref==3.6.0
Django==4.2
django-prometheus==2.2.0
djangorestframework==3.14.0
mysqlclient==2.1.1
prometheus-client==0.16.0
Pygments==2.14.0
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
